Remove ruby version requirement from Gemfile
As this version is different all over, at least on the patch level,
we'll just remove it, and hope that folks have a reasonable recent
ruby version installed.

For deploying to production, we'll figure out how to specify that
version some other way. .ruby_version should work on Heroku-like
platforms.