Sets background color of navigation bar to dark gray.
Swapping colors for the links on the top nav bar based on what type of page the navigation bar is shown was becoming messy if the navigation bar is supposed to be reused and inserted as a partial. This keeps things consistent until we do a proper overhaul of how to handle the navigation and all its styles.