Delete circle.yml
It did not work.

## Explanation

We've set up https://circleci.com/gh/iterate/iterate.github.com as a CI. But our build fails because of the way CircleCi detects ruby versions.

CircleCi reads the Gemfile to find the ruby version. But a Gemfile is actually Ruby code that allows for logic like the ```ruby ENV[RUBVY_VERSION]||2.2.2``` we have in ours. CircleCi appearantly just looks for a line starting with ```ruby``` and tries to *parse* ```ENV[RUBVY_VERSION]||2.2.2``` as a Ruby version rather than *evaluating* it as a line of Ruby code.

We can set the Ruby version CircleCi uses excplicitly in ```circle.yml```, but this still does not work because even though CircleCi picks up the right version, somehow it does not match the ```2.2.2``` Bundler thinks we are using.

Since CircleCi passing appearantly is not a big deal, I'm deleting this file for now.